Refactoring contact monitor as a library
